A-Packages: Organize classes into packages
* Moved app runner classes (SonOfAnton, Echo, EchoSession) to com.ip.arshelle.app
* Moved Ui to com.ip.arshelle.ui
* Moved Storage to com.ip.arshelle.storage
* Moved Parser to com.ip.arshelle.parser
* Moved Command interface and command classes (AddTodo, AddDeadline, AddEvent, List, Mark, Unmark, Delete, Bye) to com.ip.arshelle.command
* Moved Task, TaskList, ToDo, Deadline, Event to com.ip.arshelle.task
* Kept all exceptions under com.ip.arshelle.exceptions
* Updated package declarations and imports to match new structure
