+++ data/duke.txt
@@ -1,4 +1,3 @@
-T | 0 | read book
 D | 0 | read book | 2025-09-06
 D | 0 | read book | 2025-10-23
 E | 0 | reading | 2025-10-21 | 2025-10-27
