+++ src/main/java/com/ip/arshelle/command/Command.java
@@ -1,5 +1,8 @@
-package com.ip.arshelle;
+package com.ip.arshelle.command;
 
+import com.ip.arshelle.storage.Storage;
+import com.ip.arshelle.task.TaskList;
+import com.ip.arshelle.ui.Ui;
 import com.ip.arshelle.exceptions.SonOfAntonException;
 
 public interface Command {
