+++ src/main/java/com/ip/arshelle/command/DeleteCommand.java
@@ -1,4 +1,9 @@
-package com.ip.arshelle;
+package com.ip.arshelle.command;
+
+import com.ip.arshelle.storage.Storage;
+import com.ip.arshelle.task.Task;
+import com.ip.arshelle.task.TaskList;
+import com.ip.arshelle.ui.Ui;
 
 public class DeleteCommand implements Command {
     private final int indexOneBased;
