+++ src/main/java/com/ip/arshelle/command/ListCommand.java
@@ -1,4 +1,8 @@
-package com.ip.arshelle;
+package com.ip.arshelle.command;
+
+import com.ip.arshelle.storage.Storage;
+import com.ip.arshelle.task.TaskList;
+import com.ip.arshelle.ui.Ui;
 
 public class ListCommand implements Command {
     @Override
