+++ src/main/java/com/ip/arshelle/task/Deadline.java
@@ -1,9 +1,8 @@
-package com.ip.arshelle;
+package com.ip.arshelle.task;
 
 import com.ip.arshelle.exceptions.InvalidDateFormatException;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
