+++ src/main/java/com/ip/arshelle/task/Event.java
@@ -1,4 +1,4 @@
-package com.ip.arshelle;
+package com.ip.arshelle.task;
 
 import com.ip.arshelle.exceptions.InvalidDateFormatException;
 import com.ip.arshelle.exceptions.SonOfAntonException;
